Animate the box tiles flipping over when they are added to the pathway and flipping back when they are removed from the pathway.
Create a parent for positioning the box tiles in the grid so that the positioning transforms in the css don’t conflict with the transforms in the animation. Use css animation instead of css transitions because properties are introduced and removed through the course of the animation, which means a transition wouldn’t be appropriate.