Remove animations based on layout. Remove the perspective shift that created an angled top view.
I first removed the css animations that required the layout to be recalculated. In the process, I followed a tutorial for the card flip effect that introduced perspective to the parent element. Then I found I could create an angled top view (see last commit). The angled top view caused the card flip to flatten into an elevator-door movement so I removed the angled top view.